+++ src/Handler/CurlMultiHandler.php
@@ -6,9 +6,11 @@ namespace GuzzleHttp\Handler;
 
 use Closure;
 use GuzzleHttp\Exception\HandlerClosedException;
+use GuzzleHttp\Exception\NetworkTimeoutException;
 use GuzzleHttp\Promise as P;
 use GuzzleHttp\Promise\Promise;
 use GuzzleHttp\Promise\PromiseInterface;
+use GuzzleHttp\Psr7\Exception\TimeoutException;
 use GuzzleHttp\TransportSharing;
 use GuzzleHttp\Utils;
 use Psr\Http\Message\RequestInterface;
@@ -138,7 +140,17 @@ final class CurlMultiHandler
     {
         $this->assertOpen();
 
-        $easy = $this->factory->create($request, $options);
+        try {
+            $easy = $this->factory->create($request, $options);
+        } catch (TimeoutException $e) {
+            /** @var PromiseInterface<ResponseInterface, mixed> */
+            return P\Create::rejectionFor(new NetworkTimeoutException(
+                'The cURL handler timed out while transferring the request body',
+                $request,
+                $e
+            ));
+        }
+
         $id = (int) $easy->handle;
 
         /** @var Promise<ResponseInterface, mixed> $promise */